requirements.txt
Flask==3.0.3
highrise-bot-sdk>=24.1.0
aiohttp==3.10.11
yarl==1.15.2
gunicorn==23.0.0
nest-asyncio==1.6.0
